Remove docking controls binding
We have the "DEL" key binded to options button
(long click), so we can go to docking mode easily.
And when in docking mode, normal bindings (pitch, yaw and roll) are mapped correctly.

This solves the problem that we cannot pilot a plane while breaking (because break is assigned to Right Grip + B)